myrmx-backend/backend/src/requirements.txt
apns2==0.7.1
boto3==1.14.55
botocore==1.17.55
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
cryptography==3.1
Django==2.2.15
django-boolean-switch==0.2.9
django-bootstrap-breadcrumbs==0.9.2
django-ckeditor==5.9.0
django-cors-headers==3.4.0
django-crispy-forms==1.9.1
django-crontab==0.7.1
django-embed-video==1.3.2
django-filter==2.4.0
django-js-asset==1.2.2
django-multiselectfield==0.1.12
django-postman==3.7
django-push-notifications==2.0.0
django-rest-auth==0.9.5
djangorestframework==3.11.1
docutils==0.15.2
drf-extensions==0.6.0
Faker==4.1.1
h2==2.6.2
hpack==3.0.0
http-ece==1.1.0
hyper==0.7.0
hyperframe==3.2.0
idna==2.10
jmespath==0.10.0
mysqlclient==2.0.0
Pillow==7.1.2
py-vapid==1.7.1
pycparser==2.20
PyJWT==1.7.1
python-dateutil==2.8.1
pytz==2020.1
pywebpush==1.11.0
PyYAML==5.3
requests==2.23.0
s3transfer==0.3.3
six==1.15.0
sqlparse==0.3.1
text-unidecode==1.3
uritemplate==3.0.1
urllib3==1.25.10
